Car property - approved

+++ src/menu/sub_menus/CarRegistryMenu.java
@@ -80,18 +80,26 @@ public class CarRegistryMenu extends CarPropertiesMenu implements DBStandardQuer
 
     @Override
     public void deleteFromTable(DB_QueryEditingHandler editingHandler,  DB_QueryRequestHandler requestHandler, UI ui) {
+
         showTable(requestHandler);
-        String query = "DELETE FROM car_registry " +
-                "WHERE " + ui.chooseWhereOptions(
-                db_dependencies.TABLE_NAMES[3], db_dependencies.CAR_REGISTRY_COLUMNS,
-                requestHandler) + ";";
+        System.out.println("Please enter ID for car to delete");
+        int carID = ui.readInteger();
+        String checkQuery = "SELECT * FROM " + db_dependencies.TABLE_NAMES[2];
+        if (requestHandler.checkIfExists(checkQuery, carID, 4)) {
+            System.out.println("The car is no longer possible to delete, it has been rented out. If " +
+                    "no longer possible to rent the car, please edit the \"avalability\" property for the car\n");
+        } else {
 
+            String query = "DELETE FROM car_registry " +
+                    "WHERE " + ui.chooseWhereOptions(
+                    db_dependencies.TABLE_NAMES[3], db_dependencies.CAR_REGISTRY_COLUMNS,
+                    requestHandler) + ";";
 
-
-        System.out.println("Now we need to do the same for Car Properties associated with the deleted car");
-        System.out.print("Please enter car_registry_id you just deleted: ");
-        editingHandler.insertQuery(query);
-        super.deleteFromTable(editingHandler, requestHandler,ui);
-    }
+            System.out.println("Now we need to do the same for Car Properties associated with the deleted car");
+            System.out.print("Please enter car_registry_id you just deleted: ");
+            editingHandler.insertQuery(query);
+            super.deleteFromTable(editingHandler, requestHandler, ui);
+        } // End of if-else statmenet
+    } // End of method
 
 }
